Implement parsing expressions
This is a big, yet simple, change. It virtually completes the body of
the parser.

We add a precedence table, which maps infix tokens to the precedence
order that we’ve defined. This defines the precedence of `+` and `-` are
the same. They’re also lower precedence than `*` and `/`, for example.

We then register a new `parseInfixExpression` function to handle these
tokens. This is used in combination with the addition to
`parseExpression`.

If the peeked token is infix, then we need to call the new
`parseInfixExpression`, passing in the current expression as the left
expression of the infix operator.

Otherwise, we handle as before.