Keep track of the emitted instructions
The test was failing with an extra OpPop. We address this by keeping
track of emitted instructions and removing any extra `OpPop`
instructions after compiling the `*ast.IfExpression` `Condition`.

This makes the test fail by having the right number and type of
instructions, but the wrong value for the `OpJumpNotTruthy` target.