Make compiler test for conditionals pass
This adds some functions for replacing instructions of the same length.
That's quite a strong assumption at this point. We are only replacing a
`OpJumpNotTruthy` with another instruction of the same type, but a
different operand (or target/instruction pointer).

Potentially a stronger name of `updateJumpTarget` would reify that
assumption.

A different IR that's type-safe and independent of the byte-size of the
encoded instructions might be worth considering if the compiler and
instructions become more complex.