Add compilation scopes
This is needed for compiling functions. We need to keep the instructions
of a function literal distinct from the instructions of the main
program.

To help with that, we define a CompilationScope which encapsulates the
concept of the instructions for the current scope, along with the
previous and last emitted instructions so that we can still do the
rewriting introduced in 238518225c873b98118d4c658902eeca61d14600.

We can then use a stack of CompilationScopes as we compile the program.