Add first draft of lexer in Rust
* Add sketch for token definitions
* Add failing test for lexing

This is my second interpreter in Rust. I did a feverish one in a
frenzied few hours coding when recovering from a virus during COVID-19.
No, it wasn't *the* virus.

That one was to interpret the Honeycomb expression language but control
the environment that it evaluates against, rather than however Honeycomb
generate their random example inputs.

That came out OK-ish, but I decided to implement Monkey in Rust to get
a better comparison for how I did the Go version, and then I'll probably
redo the Honeycomb one.