Add support for parsing let statements
This is the minimal slice I could think of to support this. A Program is
a series of Statements. The first type of Statement that we support is
`let`.

Parts of Rust starting to show through. I've decided that parsing can
fail (for example, with a syntactically invalid program). So we define
an alias for `Result` and start using that. To support that usage, we
define `ParserError` to report when we see problems.