Add AST changes for Hash Literal
This was more involved than expected. Hash support for a language is
interesting.

Java has well-defined contracts/expectations for hashCode and equals(),
but generics there use type-erasure, so it all compiles down to Object.

Go enforces what can go in a map.

Rust similarly requires traits to be implemented. Trying this for now,
but I'm not sure whether this design will work.

I've added indexmap as a dependency, so that we have stable iteration
order for things. This will makes tests easier. It's not a design choice
we'd want in a proper language though.