Fix environment references
I was wondering what a benchmark of a Monkey Fibonacci function would
perform like. I discoved a bug which meant it didn't support a recursive
definition of Fibonacci.

The Environment extension did not retain a reference to the outer
Environment. Instead, it had a copy. This is a problem. If you have a
function definition like:

    let fibonacci = fn(x) {
        if (x == 0) {
            0
        } else {
            if (x == 1) {
                return 1;
            } else {
                fibonacci(x - 1) + fibonacci(x - 2);
            }
        }
    };
    fibonacci(10);

We will have 2 Environments being used here. The initial outer one will
not have a definition of the identifier 'fibonacci' until the 'let'
statement is evaluated. The inner environment created for the function
needs to have a reference (not a copy!) of the outer one so that it is
able to dynmically resolve the 'fibonacci' identifer when the function
is called.

This meant changing the Environment to stop using `&mut` everywhere and
instead use `std::rc::Rc` and `std::cell::RefCell`.

At the moment this means Rc and RefCell leak into everything. I'm not
sure if there's a cleaner way?

See:

* https://doc.rust-lang.org/std/rc/index.html
* https://doc.rust-lang.org/std/cell/index.html
* all the pain when making https://github.com/jabley/lists/blob/master/src/fourth.rs