Add start of bytecode generation
This adds a new Compiler struct which is responsible for taking an AST
and generating a stream of bytecode instructions.

It adds a test to show how a simple integer arithmetic expression may be
processed. We've added a constant pool to our bytecode format, and a
stream of instructions.

When we get to the VM, we'll start getting in to what fetches, decodes
and executes the bytecode.