Add VM that does fetch, decode, and execute
Initial version. This is making it work first. Making it fast might come
later. We create a Vec for the stack and populate it entirely, on the
basis that indexing via a stack pointer will be faster than pushing,
popping and potential resizing.

Experimenting with unifiying error handling – see the `run_vm_test`
function which deals with CompilerError and VMError. See
https://doc.rust-lang.org/stable/rust-by-example/error/multiple_error_types/wrap_error.html
for more details on this approach.

The API feels like it needs work, but will do for now. Ownership and
lifetimes will hopefully emerge as it becomes more fully featured.