Use a macro to generate Op enum
Trying out a solution for maintaining a mapping between bytes and enum.
Not sure this is the Rustacean approach, but it's what I have right now.

Tried to also add generating the `name` and other functions. Failed to
get macro generation to do what I wanted. I couldn't get it to generate
the arms in a match expression. I also couldn't get it to reference
`self` in the generated code. Feels like the macro system is purposely
with limited power?