Add support for compiling if expressions
Our grammar is:

```
if (conditional) {
    consequence
} [else {
    alternative
}]

}
```
Our first implementation is based on adding a couple of new
instructions:

* JumpNotTruthy – used to dispatch to after the consequence
* JumpTruthy – used to dispatch to the consequence

We also add the capability to track emitted instructions in the
compiler. We are emitting a dummy target for the jump initially, because
we don't know how many instructions will be compiled from the
BlockStatement body.

Once we know how large the body is, we know the location of the target
to jump to for after the consequence, and we can update that instruction
with the correct target.

This adds some functions for replacing instructions of the same length.
That's quite a strong assumption at this point. We are only replacing a
`Op::JumpNotTruthy` with another instruction of the same type, but a
different operand (or target/instruction pointer).

Potentially a stronger name of `updateJumpTarget` would reify that
assumption.

A different IR that's type-safe and independent of the byte-size of the
encoded instructions might be worth considering if the compiler and
instructions become more complex.