Add compilation scopes
This is needed for compiling functions. We need to keep the instructions
of a function literal distinct from the instructions of the main
program.

To help with that, we define a CompilationScope which encapsulates the
concept of the instructions for the current scope, along with the
previous and last emitted instructions so that we can still do the
rewriting introduced in a26d9ad999049f550913d0881cbc0c9c687e1c7c.

We can then use a stack of CompilationScopes as we compile the program.