Add Frames to the VM
This introduces activation records, or call frames, to the entire VM.
We will use this as a coherent approach for treating everything (even
main) as a function call.

Removes a single set of instructions and instruction pointer for a stack
of Frames.

This is preparatory work to support function calls in the VM.