Make SymbolTable define scope-aware
This introduces a new local scope. We want symbol tables to keep track
of which scope we're in. `define` and `resolve` should then be
scope-aware.

So the tests state that we want to use SymbolTable in a recursive-lookup
fashion.

I suspect that this will need a bit of work later to get the
borrow-checker to be happy. Compare with the effort that went into the
Environment implementation. But it's a starting point. We can hopefully
keep the SymbolTable API stable, but might need to heavily modify the
internals.