Ensure compiler scopes include symbol tables
`enter_scope` and `leave_scope` need to modify the symbol table context.

Again, more ownership joys. Needed to borrow the symbol table when
enclosing it in `enter_scope`. That rippled.

Needed to `as_ref().unwrap().borrow()` in `leave_scope`. That might be
possible doing `take()` but I didn't try too hard.

And then we need to `clone()`, so that was derived. `PartialEq` was
added so that we could test that SymbolTables were equal or not. And
Debug was also derived so that we could get meaningful error messages
when making assertions about SymbolTables.