Add local binding support to the VM
We've chosen to store local bindings to a function on the stack, rather
than in a separate data structure.

This is consistent with the current approach of using the stack for
everything. It's also useful to learning more about how to do this,
since it's a common thing in comp-sci and low-level programming.

It works like this way:

* When the VM encounters an `OpCall` instruction, we store the current
  value of the stack pointer.
* We then bump the stack pointer by the number of locals used by the
  function we're about to execute.
  * This leaves a 'hole' in the stack.
* This hole can then be filled with the local bindings.

		|                 |
		|                 |
		+-+-+-+-+-+-+-+-+-+
		|                 |
		+-+-+-+-+-+-+-+-+-+
    vm.sp ->	|                 |
      		+-+-+-+-+-+-+-+-+-+
		|     Local 2     |  -+
		+-+-+-+-+-+-+-+-+-+   |-- reserved for locals
		|     Local 1     |  -+
		+-+-+-+-+-+-+-+-+-+
		|    Function     |
		+-+-+-+-+-+-+-+-+-+
		|  Other Value 2  |  <-+
		+-+-+-+-+-+-+-+-+-+    |-- pushed before call
		|  Other Value 1  |  <-+
		+-+-+-+-+-+-+-+-+-+

This makes post-execution of a function cleaner. Since we saved the
previous value of the stack pointer, we simpy restore that and 'reset'
the stack.

A CompiledFunction needs to know how many local bindings the function is
going to create. We can get that from the scoped symbol tables.

We enhance Frame to track the base pointer (also known as the frame
pointer in the literature).

When the VM calls a function, it needs to allocate space on the stack
for the local bindings before executing the function. This means bumping
the vm stack pointer without pushing values. OpReturnValue and OpReturn
need to have an associated change to move the stack pointer back to
where execution should resume. We have a minor optimisation here – we
move it past the location of the CompiledFunction itself, so that we no
longer need to pop that off the stack too.

OpSetLocal and OpGetLocal can then be implemented to pop and push items
onto their part of the stack.

This designs seems to have actually worked! I was worried I would have
to start using RefCell and Rc again for parts of it, but I've got away
with cloning instead. Maybe performance will be shocking due to all that
copying?