Add support for compiling built-ins
Define built-ins in the SymbolTable when we create the Compiler.

We need to add a field to keep track of the number of definitions,
because built-in definitions don't count when we're calculating the
index to assign. So we can no longer use the number of items in the
store to tell us this. Instead, we need to track it separately.

Define constant to index into the built-ins. This gives us the same
ordering as the Go implementation. My first attempt wasn't right,
because it had them in a different order (the encountered order in the
text). This was giving the wrong index for `push`. So I added fuller
error messages to highlight this issue.

Load the symbol from the appropriate scope. When we resolve an
identifer, we need to emit the right bytecode to load from the right
scope.