Add Op::Closure opcode
`Op::Closure` tells the VM to wrap the `Object::CompiledFunction` in an
`Object::Closure`.

It has 2 operands. The first (2 bytes wide) is the constant index. This
describes where we can find the `Object::CompiledFunction` that's to be
converted into a closure.

The second operand (1 byte wide) is the number of free variables for the
closure. This limits us to 256 free variables per closure. If you have a
closure that has more than 256 free variables, you might be writing code
which is hard to understand :D