Make SymbolTable aware of free variables
Every non-local, non-global, non-built-in is treated as a free variable.

See https://en.wikipedia.org/wiki/Free_variables_and_bound_variables

This was interesting to see how mutability permeates through the design
as is, and ripples when it's changed.

If `resolve` returns a variable in an outer scope which is not a global
or built-in, then treat it as a free variable.

We have now changed `resolve` from only borrowing `self` to requiring
`&mut`. Calling `resolve` may mutate the SymbolTable.