Add new binaries
Thus far, I'm unable to run criterion benchmarks with flamegraph. I did
describe a way of generating flamegraphs, but it's a bit nasty. See
835d8f8d.

So intead, I've added a couple of new binaries.

Now, I can do:

   cargo build --release

and then

   sudo cargo flamegraph --bin vm-flamegraph

or

    sudo cargo flamegraph --bin interpreter-flamegraph

I need to use `sudo` because I'm on MacOS and using `dtrace` with

    dtrace: system integrity protection is on, some features will not be available

But it generates me a flamegraph!