Extract all op dispatch into functions
Hoping this will make profiling and performance improvement easier if we
can see where time is being spent. At the moment, the flamegraph shows
99% of time is spent in `vm::run`. Need more fine-grained view to see if
there are any other candidates for optimising.

This doesn't seem to affect benchmarks. Presumably the compiler is smart
enough to inline, or jump is very cheap.